refactor: community 타입 파일로의 이동 및 관련 import 경로 수정
