+++ src/pages/CommunityList/_components/PostItem.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/native';
 import React from 'react';
 
 import { CATEGORY_LABEL } from '@/store/community.mock';
-import type { Post } from '@/types/community';
+import type { Post } from '@/types/community.type';
 
 interface Props {
   post: Post;
