+++ src/pages/CommunityList/index.tsx
@@ -7,7 +7,7 @@ import { FlatList, View } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import { samplePosts } from '@/store/community.mock';
-import type { CategoryKey } from '@/types/community';
+import type { CategoryKey } from '@/types/community.type';
 import type {
   RootStackParamList,
   TabParamList,
