+++ src/types/user.types.ts
@@ -1,0 +1,7 @@
+export type User = {
+  id: number;
+  email: string;
+  nickname: string;
+  avatarUrl: string;
+  createdAt: string;
+};
